fix: The composable way for app screen building
Since Library screen builds a scaffold and the other screens probably
will to. The whole app is inside a Box composable so we need to only
show our bottom bar for the pages that need them. And since we are using
one nav controller we can navigate to any page from any page and
build/design the page as we see fit.